Add router.resolve to abstract source to url logic
