+++ src/utilities/createRouterPush.spec.ts
@@ -1,5 +1,6 @@
 import { expect, test, vi } from 'vitest'
 import { createRouterPush } from '@/utilities/createRouterPush'
+import { createRouterResolve } from '@/utilities/createRouterResolve'
 import { resolveRoutes } from '@/utilities/resolveRoutes'
 import { createRouterNavigation } from '@/utilities/routerNavigation'
 import { routes } from '@/utilities/testHelpers'
@@ -8,7 +9,8 @@ test('push calls onLocationUpdated', () => {
   const onLocationUpdate = vi.fn()
   const navigation = createRouterNavigation({ onLocationUpdate })
   const resolved = resolveRoutes(routes)
-  const push = createRouterPush<typeof routes>({ navigation, resolved })
+  const resolve = createRouterResolve({ resolved })
+  const push = createRouterPush<typeof routes>({ navigation, resolve })
 
   push({ route: 'parentA', params: { paramA: '' } })
